refactor(engine): extract cycle orchestration into CycleOrchestrationMixin (phase 5)
Splits the other god-ish file the review flagged. The trading-cycle drivers
(run_cycle and its starved / strategic / allocated variants), candidate
execution, and the price-history helper move verbatim into
auramaur/strategy/engine_cycle.py as CycleOrchestrationMixin, mixed into
TradingEngine. Pure structural move — the cycle methods still call the engine's
analysis / scan / risk / gateway / allocator via self, and the bot drives
run_cycle via self.

The AST sweep caught two runtime symbols the test paths don't exercise:
CapitalAllocator.compute_expected_value(...) and CandidateTrade(...) construction
in the strategic/allocated cycles — both now imported (would have NameError'd on
those branches). TradeCandidate is annotation-only (TYPE_CHECKING). Also repaired
an extraction off-by-one that had truncated _get_price_history's final return.

engine.py: 1504 -> 750 lines (-50%). Parity proven by the existing engine /
cycle / allocator suites. Full suite: 837 passed.
